RIDER-132995 add webapps icons fro the nodes in the tree
