+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/resources/META-INF/azure-intellij-plugin-appservice-dotnet.xml
@@ -105,6 +105,8 @@
 
         <xdebugger.dialog.process.view.provider implementation="com.microsoft.azure.toolkit.intellij.debugger.view.AttachToAzureAppServiceProcessViewProvider" />
         <xdebugger.attachDebuggerProvider implementation="com.microsoft.azure.toolkit.intellij.debugger.RiderAzureAttachDebuggerProvider" />
+
+        <iconMapper mappingFile="AppServiceIconMappings.json"/>
     </extensions>
 
     <actions>
